backend/functions/requirements.txt
firebase-admin>=6.2.0
firebase-functions>=0.1.0
google-genai>=0.2.0
